feat(UI/UX): Moved dashboard from root to /dashboard and put uploads into a seperate component
