+++ hansroslinger/src/constants/application.ts
@@ -1,0 +1,1 @@
+export const ACCEPTED_FILE_TYPES = ["png, json"]
