+++ hansroslinger/src/types/application.d.ts
@@ -1,0 +1,10 @@
+import { ACCEPTED_FILE_TYPES } from "constants/application"
+
+// This will change once we know what we are displaying
+// In future milestone will change to a type generated by the backend
+export type Uploads = {
+  name: string,
+  type: (typeof ACCEPTED_FILE_TYPES)[number],
+  src: string
+  imageSrc: string
+}
